pyproject.toml
[project]
name = "aaidc-module1-genbotx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.15",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langgraph>=0.6.3",
    "loguru>=0.7.3",
    "ollama>=0.5.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "streamlit>=1.48.0",
    "wikipedia>=1.4.0",
    "wikipedia-api>=0.8.1",
]
